Rename the GPU timing hook and drop its vestigial cupy guard
The module implements only GPUTimings now, so log_timings.py no longer
described it. Renamed to log_GPU_timings.py; both importers updated.

The try/except around 'import cupy' is also removed. It was load-bearing
only while this module still held CPUTimings: core/controller.py imports
that on every run, so a bare 'import cupy' would have crashed every
CPU-only run. After the split nothing on a CPU path reaches this module -
both importers are function-local and behind 'if args[useGPU]' - so cupy
is present whenever it is loaded. Checked that nothing else pulls it in
without cupy: sphinx already mocks cupy via autodoc_mock_imports, no hook
auto-discovery exists, and the package-walking helpers in
tests/test_helpers/test_helpers.py are uncalled (that file defines no tests).

Removing the guard also improves diagnostics: 'cp = None' turned a missing
dependency into an AttributeError on NoneType inside a timing call, where a
plain import raises ModuleNotFoundError naming cupy at the point of use.

Note the guard in projects/RayleighBenard/benchmarks/print_timings_hook.py
is still needed - that module continues to hold both PrintCPUTimings and
PrintGPUTimings, and RBC3D_configs.py imports the two together.
